+++ e2e/integration/utils/utils.js
@@ -29,22 +29,3 @@ module.exports.waitUntil = async function(action, timeout = 15000, delayDuration
 };
 
 module.exports.delay = delay;
-
-const createManifestForJPadKey = key_path => ({
-  key_path: `${key_path}`,
-  meta: {
-    name: 'aaaaaaa',
-    tags: [],
-    description: '',
-    archived: false,
-  },
-  implementation: {
-    type: 'file',
-    format: 'jpad',
-  },
-  valueType: 'number',
-  dependencies: [],
-  enabled: true,
-});
-
-module.exports.createManifestForJPadKey = createManifestForJPadKey;
